requirements.txt
Flask
rembg[cpu]
Pillow
onnxruntime
